Put host tag back
I didn't realize that I needed to be running the server so that it would be reachable on other devices on the network. I learned that 0.0.0.0 simultaneously runs localhost and the computers ip address to accept requests coming from other devices. This also included fixing my fetch request to the backend to be the "servers" (my computer) ip, not localhost